Updated buttons
Added the rest of the buttons so that there is acts 1-10. Also placed all of the buttons into a main menu class and changed the window to used stacked widgets. This is so that later the window can be easily cleared and re-called when needed without having to use multiple windows